fastapi/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
torch==2.2.1
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
visdom==0.2.4
websocket-client==1.7.0
